Fix Rust compilation errors: E0382 moved reader and callback
- Fix E0382 in lsp_transport/transport.rs:221: Use read_line() instead of lines() iterator to prevent reader from being consumed in loop
  - Changed from 'for line in reader.lines()' to manual 'read_line()' loop
  - Allows reader to be used for read_exact() after header parsing

- Fix E0382 in chat_sidebar/rag_chat.rs:252: Wrap FnMut callback in Arc<Mutex>
  - Avoids moving callback into closure while still calling it after
  - Lock mutex before calling callback in both streaming and final chunk

- Add tokio::sync::Mutex import to chat_sidebar/rag_chat.rs

22 errors remaining (down from 24)

+++ src-tauri/src/agent_editor/planner.rs
@@ -177,7 +177,7 @@ impl EditPlanner {
 
                 let end_line = edit.get("end_line")
                     .and_then(|e| e.as_i64())
-                    .unwrap_or(start_line) as usize;
+                    .unwrap_or(start_line as i64) as usize;
 
                 let new_content = edit.get("new_content")
                     .or_else(|| edit.get("content"))
